+++ src/constants/Extension.ts
@@ -18,7 +18,6 @@ export const COMMAND_NAME = {
   createCategory: getCommandName("createCategory"),
   exportTaxonomy: getCommandName("exportTaxonomy"),
   remap: getCommandName("remap"),
-  setDate: getCommandName("setDate"),
   setLastModifiedDate: getCommandName("setLastModifiedDate"),
   generateSlug: getCommandName("generateSlug"),
   createFromTemplate: getCommandName("createFromTemplate"),