+++ src/constants/settings.ts
@@ -5,9 +5,10 @@ export const CONFIG_KEY = "frontMatter";
 export const SETTING_TAXONOMY_TAGS = "taxonomy.tags";
 export const SETTING_TAXONOMY_CATEGORIES = "taxonomy.categories";
 export const SETTING_DATE_FORMAT = "taxonomy.dateFormat";
+export const SETTING_COMMA_SEPARATED_FIELDS = "taxonomy.commaSeparatedFields";
+export const SETTING_TAXONOMY_CONTENT_TYPES = "taxonomy.contentTypes";
 export const SETTING_DATE_FIELD = "taxonomy.dateField";
 export const SETTING_MODIFIED_FIELD = "taxonomy.modifiedField";
-export const SETTING_COMMA_SEPARATED_FIELDS = "taxonomy.commaSeparatedFields";
 
 export const SETTING_SLUG_PREFIX = "taxonomy.slugPrefix";
 export const SETTING_SLUG_SUFFIX = "taxonomy.slugSuffix";