+++ src/viewpanel/components/Actions.tsx
@@ -2,9 +2,7 @@ import * as React from 'react';
 import { PanelSettings } from '../../models/PanelSettings';
 import { Collapsible } from './Collapsible';
 import { CustomScript } from './CustomScript';
-import { DateAction } from './DateAction';
 import { Preview } from './Preview';
-import { PublishAction } from './PublishAction';
 import { SlugAction } from './SlugAction';
 
 export interface IActionsProps {
@@ -26,10 +24,6 @@ export const Actions: React.FunctionComponent<IActionsProps> = (props: React.Pro
         { metadata && metadata.title && <SlugAction value={metadata.title} crntValue={metadata.slug} slugOpts={settings.slug} /> }
 
         { settings?.preview?.host && <Preview slug={metadata.slug} /> }
-
-        <DateAction />
-
-        { metadata && typeof metadata.draft !== undefined && <PublishAction draft={metadata.draft} />}
 
         {
           (settings && settings.scripts && settings.scripts.length > 0) && (