+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [#101](https://github.com/estruyf/vscode-front-matter/issues/101): Date picker available on the metadata section
 - [#102](https://github.com/estruyf/vscode-front-matter/issues/102): Support comma separated arrays in front matter
 - [#103](https://github.com/estruyf/vscode-front-matter/issues/103): Added title and description field to the metadata section
+- [#104](https://github.com/estruyf/vscode-front-matter/issues/104): Allow to set images in front matter from the metadata panel section
+- [#105](https://github.com/estruyf/vscode-front-matter/issues/105): Content Type support with backwards compatibility
 
 ## [3.1.0] - 2021-09-10